+++ config.ts
@@ -1,0 +1,17 @@
+import { GasPrice } from "npm:@cosmjs/stargate";
+
+export interface Config {
+  // dapp chain
+  endpoint: string;
+  monitoringContract: string;
+  proxyContract: string;
+  mnemonic: string;
+  feeDenom: string;
+  addressPrefix: string;
+  chainId: string;
+  gasPrice: GasPrice;
+
+  // nois chain (Testnet 005)
+  noisEndpoint: string;
+  drandContract: string;
+}
