+++ docker-compose.yml
@@ -1,7 +1,21 @@
 version: '3'
 
 services:
-  nois-pingpong:
-    image: docker.io/noislabs/nois-pingpong:latest
+  nois-pingpong-testnet:
+    build:
+      context: .
+      dockerfile: Dockerfile
+      args:
+        CONFIG_FILE: config/testnet.ts
+    image: docker.io/noislabs/nois-pingpong:testnet
     ports:
       - "19000:3001"
+  nois-pingpong-juno-public:
+    build:
+      context: .
+      dockerfile: Dockerfile
+      args:
+        CONFIG_FILE: config/juno-public.ts
+    image: docker.io/noislabs/nois-pingpong:juno-public
+    ports:
+      - "19001:3001"
